Retrofit, ViewModel, Activity 수정
