+++ app/src/main/java/com/example/myapp/MapActivity.kt
@@ -1,11 +1,13 @@
 package com.example.myapp
 
 import android.os.Bundle
+import android.util.Log
 import android.webkit.WebView
 import android.webkit.WebViewClient
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.viewModels
+import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
@@ -13,8 +15,10 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.livedata.observeAsState
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
 import androidx.compose.ui.viewinterop.AndroidView
 import com.example.myapp.stage.AddressViewModel
 
@@ -24,46 +28,54 @@ class MapActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
 
+
+        val targetId = intent.getIntExtra("stageId", -1)
+        if (targetId != -1) {
+            addressViewModel.getAddresses(targetId)
+        } else {
+            Log.e("MapActivity", "Invalid stageId received")
+        }
+
         setContent {
             MapScreen()
         }
-
-        val page: Int = 0
-        val size: Int = 500
-        addressViewModel.fetchAddresses(page, size)
 
     }
 
     @Composable
     fun MapScreen() {
-        val responseData by addressViewModel.responseData.observeAsState()
+        val stageData by addressViewModel.responseData.observeAsState()
         val errorMessage by addressViewModel.errorMessage.observeAsState()
+        val isLoaded = addressViewModel.isStageDataLoaded.value
 
         Column(modifier = Modifier.padding(16.dp)) {
             errorMessage?.let {
                 Text(text = it)
             }
 
-            responseData?.let { data ->
-                val targetId = intent.getIntExtra("stageId", -1)
-                val content = data.content.find { it.indieStreetId == targetId }
-                content?.let { addr ->
-                    val placeLink = addr.placeLink
-                    if (placeLink.isBlank()) {
-                        Text(text = "address NOT FOUND")
-                    }
-                    AndroidView(factory = {
-                        WebView(it).apply {
-                            webViewClient = WebViewClient()
-                            settings.javaScriptEnabled = true
-                            loadUrl(placeLink)
-                        }
-                    }, modifier = Modifier.fillMaxSize())
-                } ?: run {
-                    Text(text = "No address found", modifier = Modifier.fillMaxSize())
+            if (!isLoaded) {
+                Box(
+                    modifier = Modifier.fillMaxSize(),
+                    contentAlignment = Alignment.Center
+                ) {
+                    Text("Now Loading...", fontSize = 25.sp)
                 }
-            } ?: run {
-                Text(text = "Loading...", modifier = Modifier.fillMaxSize())
+                return
+            }
+
+            stageData?.let { data ->
+                val placeLink = data.placeLink
+                if (placeLink.isBlank()) {
+                    Text(text = "address NOT FOUND")
+                    return
+                }
+                AndroidView(factory = {
+                    WebView(it).apply {
+                        webViewClient = WebViewClient()
+                        settings.javaScriptEnabled = true
+                        loadUrl(placeLink)
+                    }
+                }, modifier = Modifier.fillMaxSize())
             }
         }
     }
