--- app/src/main/java/com/example/myapp/Retrofit/StageResponse.kt
@@ -1,15 +1,0 @@
-package com.example.myapp.Retrofit
-
-data class StageResponse(
-    val totalPages: Int,
-    val totalElements: Int,
-    val size: Int,
-    val content: List<StageData>,
-    val number: Int,
-    val sort: SortInfo,
-    val first: Boolean,
-    val last: Boolean,
-    val numberOfElements: Int,
-    val pageable: PageableInfo,
-    val empty: Boolean
-)
