+++ app/src/main/java/com/example/myapp/stage/AddressViewModel.kt
@@ -1,32 +1,36 @@
 package com.example.myapp.stage
 
 import android.util.Log
+import androidx.compose.runtime.mutableStateOf
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.example.myapp.Retrofit.LiveResponse
-import com.example.myapp.Retrofit.RetrofitFactory
 import com.example.myapp.Retrofit.RetrofitFactory.addressService
-import com.example.myapp.Retrofit.StageResponse
+import com.example.myapp.Retrofit.StageData
 import kotlinx.coroutines.launch
 
 class AddressViewModel : ViewModel() {
-    private val _responseData = MutableLiveData<StageResponse?>()
-    val responseData: LiveData<StageResponse?> get() = _responseData
+    private val _responseData = MutableLiveData<StageData?>()
+    val responseData: LiveData<StageData?> get() = _responseData
 
     private val _errorMessage = MutableLiveData<String?>()
     val errorMessage: LiveData<String?> get() = _errorMessage
 
-    fun fetchAddresses(page: Int = 0, size: Int = 500) {
+    val isStageDataLoaded = mutableStateOf(false)
+
+    fun getAddresses(indieStreetId: Int) {
+        isStageDataLoaded.value = false
         viewModelScope.launch {
             try {
-                val responseData = addressService.getAddressById(page, size)
+                Log.d("AddressViewModel", "id: $indieStreetId")
+                val responseData = addressService.getAddressById(indieStreetId)
                 _responseData.value = responseData
-                Log.d("AddressViewModel", "Fetched response data: $responseData")
+                Log.d("AddressViewModel", "Response data: $responseData")
+                isStageDataLoaded.value = true
             } catch (e: Exception) {
                 _errorMessage.value = "Error: ${e.message}"
-                Log.e("AddressViewModel", "Failed to fetch addresses", e)
+                Log.e("AddressViewModel", "Failed to get addresses", e)
             }
         }
     }
