+++ app/src/main/java/com/example/myapp/stage/StageAddressService.kt
@@ -1,16 +1,12 @@
 package com.example.myapp.stage
 
-import com.example.myapp.Retrofit.LiveResponse
-import com.example.myapp.Retrofit.StageResponse
-import retrofit2.Call
+import com.example.myapp.Retrofit.StageData
 import retrofit2.http.GET
 import retrofit2.http.Query
 
 interface StageAddressService {
-    @GET("/stages")
+    @GET("/stage")
     suspend fun getAddressById(
-        @Query("page") page: Int = 0,
-        @Query("size") size: Int = 10
-    ): StageResponse
-
+        @Query("indieStreetId") id: Int
+    ): StageData
 }
